mds-703 Split python database model files (#239)
* Split location model class file

* Split mine model class file

* Split mine status model class file

* Split permit model class file

* Split party model class file
