+++ python-backend/app/api/mines/mine/resources/mine.py
@@ -5,11 +5,14 @@ from flask import request
 from flask_restplus import Resource, reqparse, inputs
 from sqlalchemy_filters import apply_sort, apply_pagination
 
-from ...status.models.status import MineStatus, MineStatusXref
-from ..models.mine import MineIdentity, MineDetail, MineralTenureXref
+from ...status.models.mine_status import MineStatus
+from ...status.models.mine_status_xref import MineStatusXref
+from ..models.mine_identity import MineIdentity
+from ..models.mine_detail import MineDetail
+from ..models.mineral_tenure_xref import MineralTenureXref
 from ....permits.permit.models.permit import Permit
-from ...location.models.location import MineLocation, MineMapViewLocation 
-
+from ...location.models.mine_location import MineLocation
+from ...location.models.mine_map_view_location import MineMapViewLocation
 from ....utils.random import generate_mine_no
 from app.extensions import jwt, api
 from ....utils.resources_mixins import UserMixin, ErrorMixin
