+++ python-backend/app/api/parties/party/resources/party.py
@@ -5,8 +5,9 @@ from flask import request
 from flask_restplus import Resource, reqparse
 from sqlalchemy import or_
 
-from ....mines.mine.models.mine import MineIdentity
-from ..models.party import Party, MgrAppointment
+from ....mines.mine.models.mine_identity import MineIdentity
+from ..models.party import Party
+from ..models.mgr_appointment import MgrAppointment
 from ....constants import PARTY_STATUS_CODE
 from app.extensions import jwt, api
 from ....utils.resources_mixins import UserMixin, ErrorMixin
